User property naming changes, other small fixes
Modified a debug print statement that was referring to an undefined
object (was causing crashing), renamed "name" to "username" in the User
model and register form. Removed the route for the users index page, as
the corresponding view was deleted in an earlier commit. Changed the
register form's password field's type to "password" to prevent the typed
password from being shown.